+++ src/main/java/com/jeffpeng/jmod/JMODRepresentation.java
@@ -3,11 +3,11 @@ package com.jeffpeng.jmod;
 import java.util.ArrayList;
 import java.util.List;
 
-import org.apache.logging.log4j.LogManager;
-import org.apache.logging.log4j.Logger;
-
 import net.minecraft.item.ItemBlock;
 import net.minecraftforge.common.MinecraftForge;
+
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
 
 import com.jeffpeng.jmod.actions.AddChestLoot;
 import com.jeffpeng.jmod.actions.AddShapedRecipe;
@@ -20,13 +20,12 @@ import com.jeffpeng.jmod.actions.SetBlockProperties;
 import com.jeffpeng.jmod.crafting.AnvilHandler;
 import com.jeffpeng.jmod.crafting.DropHandler;
 import com.jeffpeng.jmod.crafting.ToolRepairRecipe;
-import com.jeffpeng.jmod.interfaces.IItem;
 import com.jeffpeng.jmod.interfaces.IBlock;
-import com.jeffpeng.jmod.interfaces.IStagedObject;
 import com.jeffpeng.jmod.interfaces.IExecutableObject;
+import com.jeffpeng.jmod.interfaces.IItem;
+import com.jeffpeng.jmod.interfaces.IStagedObject;
 import com.jeffpeng.jmod.modintegration.rotarycraft.PatchRoCSteelTools;
 import com.jeffpeng.jmod.primitives.JMODInfo;
-import com.jeffpeng.jmod.registry.PlayerData;
 import com.jeffpeng.jmod.scripting.JScript;
 import com.jeffpeng.jmod.tooltipper.ToolTipper;
 import com.jeffpeng.jmod.validator.Validator;
@@ -55,7 +54,6 @@ public class JMODRepresentation {
 	private JMODRepresentation instance = this;
 	private boolean scriptingFinished = false;
 	private boolean scriptingErrored = false;
-	private PlayerData playerData;
 	
 	protected List<IStagedObject> stageables = new ArrayList<>();