+++ src/main/java/com/jeffpeng/jmod/actions/AddShapedRecipe.java
@@ -3,7 +3,6 @@ package com.jeffpeng.jmod.actions;
 import java.util.List;
 
 import com.jeffpeng.jmod.JMODRepresentation;
-import com.jeffpeng.jmod.API.Blacklist;
 import com.jeffpeng.jmod.crafting.BlacklistCraftingResults;
 import com.jeffpeng.jmod.crafting.StringListRecipe;
 import com.jeffpeng.jmod.primitives.BasicAction;