+++ src/main/java/com/jeffpeng/jmod/modintegration/nei/RecipeHandler.java
@@ -1,17 +1,8 @@
 package com.jeffpeng.jmod.modintegration.nei;
 
-import java.util.List;
-
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.CraftingManager;
-import net.minecraft.item.crafting.IRecipe;
-import net.minecraft.item.crafting.ShapedRecipes;
-import net.minecraftforge.oredict.ShapedOreRecipe;
 import codechicken.nei.NEIServerUtils;
 import codechicken.nei.recipe.ShapedRecipeHandler;
-import codechicken.nei.recipe.ShapedRecipeHandler.CachedShapedRecipe;
-
-import com.jeffpeng.jmod.JMOD;
 import com.jeffpeng.jmod.crafting.StringListRecipe;
 
 public class RecipeHandler extends ShapedRecipeHandler {