+++ src/main/java/com/jeffpeng/jmod/util/ForgeDeepInterface.java
@@ -28,7 +28,6 @@ import cpw.mods.fml.common.LoadController;
 import cpw.mods.fml.common.Loader;
 import cpw.mods.fml.common.ModContainer;
 import cpw.mods.fml.common.registry.GameData;
-import cpw.mods.fml.common.registry.GameRegistry;
 
 public class ForgeDeepInterface {